requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
blessed==1.19.1
charset-normalizer==2.1.1
discord-ui==5.1.6
discord.py==2.1.0
enlighten==1.10.2
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.2
idna==3.4
imageio-ffmpeg==0.4.7
multidict==6.0.3
nsz==4.0.0
numpy==1.23.5
prefixed==0.3.2
pycryptodome==3.14.1
six==1.16.0
wcwidth==0.2.5
yarl==1.8.2
zstandard==0.17.0
